aha ya sede 5er 5er b3een allah

+++ libft/ft_strrchr.c
@@ -4,7 +4,7 @@ char	*ft_strrchr(const char *s, int c)
 {
 	int	i;
 
-	i = (int)ft_strlen(s);
+	i = (ft_strlen(s));
 	while (i >= 0)
 	{
 		if (s[i] == (char)c)
